+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.1.1 Under development
 
+- Bug #6: Remove generate id in `UIAwesome\Html\FormControl\Button` class (@terabytesoftw)
+
 ## 0.1.0 March 6, 2024
 
 - Initial release.